Moved types and functions related to module content from common_storage.ts to module_content.ts.
