+++ src/blocks/mrc_mechanism.ts
@@ -28,8 +28,8 @@ import { Editor } from '../editor/editor';
 import { ExtendedPythonGenerator } from '../editor/extended_python_generator';
 import { getAllowedTypesForSetCheck } from './utils/python';
 import * as toolboxItems from '../toolbox/items';
-import * as commonStorage from '../storage/common_storage';
 import * as storageModule from '../storage/module';
+import * as storageModuleContent from '../storage/module_content';
 import * as storageNames from '../storage/names';
 import * as value from './utils/value';
 import { renameMethodCallers } from './mrc_call_python_function'
@@ -153,7 +153,7 @@ const MECHANISM = {
     }
     return legalName;
   },
-  getMechanism: function (this: MechanismBlock): commonStorage.MechanismInRobot | null {
+  getMechanism: function (this: MechanismBlock): storageModuleContent.MechanismInRobot | null {
     const mechanismName = this.getFieldValue(FIELD_NAME);
     const mechanismType = this.mrcImportModule + '.' + this.getFieldValue(FIELD_TYPE);
     return {
@@ -174,7 +174,7 @@ const MECHANISM = {
       // Then here, we'd look for the mechanism with the marching UUID, and we'd update the
       // FIELD_TYPE value if the mechanism's class name had changed.
       let foundMechanism: storageModule.Mechanism | null = null;
-      const components: commonStorage.Component[] = []
+      const components: storageModuleContent.Component[] = []
       for (const mechanism of editor.getMechanisms()) {
         if (mechanism.className === this.getFieldValue(FIELD_TYPE)) {
           foundMechanism = mechanism;
@@ -247,7 +247,7 @@ export const pythonFromBlock = function (
 }
 
 export function createMechanismBlock(
-    mechanism: storageModule.Mechanism, components: commonStorage.Component[]): toolboxItems.Block {
+    mechanism: storageModule.Mechanism, components: storageModuleContent.Component[]): toolboxItems.Block {
   const snakeCaseName = storageNames.pascalCaseToSnakeCase(mechanism.className);
   const mechanismName = 'my_' + snakeCaseName;
   const extraState: MechanismExtraState = {
